+++ src/render3d/gpu/Renderer3DMeshCache.hpp
@@ -26,6 +26,7 @@ namespace luax::render3d {
 
         std::unordered_map<std::uint64_t, GpuMesh> m_gpuMeshes;
         std::unordered_map<std::uint64_t, unsigned int> m_gpuTextures;
+        unsigned m_gen = 0;
     };
 
 } // namespace luax::render3d