+++ src/render3d/gpu/ViewportComposite.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace luax::render3d {
-
-    void drawViewportComposite(unsigned int colorTexture, float width, float height);
-
-} // namespace luax::render3d